+++ src/main/java/com/example/review/domain/members/service/implement/CartServiceImpl.java
@@ -26,15 +26,12 @@ import org.springframework.transaction.annotation.Transactional;
 public class CartServiceImpl implements CartService {
 
     private final CartRepository cartRepository;
-    private final MemberRepository memberRepository;
     private final ItemRepository itemRepository;
 
     //장바구니 생성
     @Override
     @Transactional
-    public CartResponse create(CreateCartRequest cartRequest, User user) {
-        Member member = getMember(user);
-
+    public CartResponse create(CreateCartRequest cartRequest, Member member) {
         Item item = existItemCheck(cartRequest.itemId());
 
         //장바구니 넣으려는 상품 수량 > 실제 상품 재고
@@ -60,16 +57,16 @@ public class CartServiceImpl implements CartService {
 
         // CartRequest.Option -> Cart.Option
         List<Cart.Option> options = cartRequest.optionValue().stream()
-                .map(option -> new Cart.Option(option.key(), option.value()))
-                .toList();
+            .map(option -> new Cart.Option(option.key(), option.value()))
+            .toList();
 
         // 엔티티는 빌더로
         Cart cart = Cart.builder()
-                .count(cartRequest.cartItemCount())
-                .member(member)
-                .item(item)
-                .optionValues(options)
-                .build();
+            .count(cartRequest.cartItemCount())
+            .member(member)
+            .item(item)
+            .optionValues(options)
+            .build();
 
         cartRepository.save(cart);
 
@@ -79,8 +76,7 @@ public class CartServiceImpl implements CartService {
     //장바구니 수정
     @Override
     @Transactional
-    public void update(Long cartId, UpdateCartRequest cartRequest, User user) {
-        Member member = getMember(user);
+    public void update(Long cartId, UpdateCartRequest cartRequest, Member member) {
         Cart cart = existMemberCartCheck(cartId, member);
 
         cart.updateCart(cartRequest.count());
@@ -90,73 +86,66 @@ public class CartServiceImpl implements CartService {
     //장바구니 전체 조회
     @Override
     @Transactional(readOnly = true)
-    public CartPageResponse getAll(Pageable pageable, User user) {
-        Member member = getMember(user);
+    public CartPageResponse getAll(Pageable pageable, Member member) {
         //회원에 해당하는 전체 장바구니
         Page<Cart> cartList = cartRepository.findAllByMember(member, pageable);
 
         List<CartResponse> cartItems = cartList.getContent().stream()
-                .map(cart -> new CartResponse(
-                        cart.getCartId(),
-                        cart.getItem().getItemId(),
-                        cart.getItem().getItemName(),
-                        cart.getItem().getItemPrice(),
-                        cart.getCount(),
-                        cart.getItemState(),
-                        cart.getItem().getItemOption().getOptionValues().stream()
-                                .map(option -> new CartResponse.Option(option.key(), option.value()))
-                                .toList()
-                ))
-                .toList();
+            .map(cart -> new CartResponse(
+                cart.getCartId(),
+                cart.getItem().getItemId(),
+                cart.getItem().getItemName(),
+                cart.getItem().getItemPrice(),
+                cart.getCount(),
+                cart.getItemState(),
+                cart.getItem().getItemOption().getOptionValues().stream()
+                    .map(option -> new CartResponse.Option(option.key(), option.value()))
+                    .toList()
+            ))
+            .toList();
 
         return new CartPageResponse(
-                cartList.getTotalPages(),
-                (int) cartList.getTotalElements(),
-                cartList.getNumber(),
-                cartList.getSize(),
-                cartItems
+            cartList.getTotalPages(),
+            (int) cartList.getTotalElements(),
+            cartList.getNumber(),
+            cartList.getSize(),
+            cartItems
         );
     }
 
     //체크된 장바구니들 삭제
     @Override
     @Transactional
-    public void deleteSelectedCarts(List<Long> cartIds, User user) {
-        Member member = getMember(user);
+    public void deleteSelectedCarts(List<Long> cartIds, Member member) {
         for (Long cartId : cartIds) {
             Cart cart = existMemberCartCheck(cartId, member);
             cartRepository.delete(cart);
         }
     }
 
-    private Member getMember(User user) {
-        return memberRepository.findByEmail(user.getUsername())
-                .orElseThrow(() -> new BusinessException(NOT_FOUND_MEMBER));
-    }
-
     //상품 존재 여부 확인
     private Item existItemCheck(Long itemId) {
         return itemRepository.findById(itemId)
-                .orElseThrow(() -> new BusinessException(NOT_FOUND_ITEM, "상품이 존재하지 않습니다."));
+            .orElseThrow(() -> new BusinessException(NOT_FOUND_ITEM, "상품이 존재하지 않습니다."));
     }
 
     //회원의 장바구니 존재 여부 확인
     private Cart existMemberCartCheck(Long cartId, Member member) {
         return cartRepository.findByCartIdAndMember(cartId, member)
-                .orElseThrow(() -> new BusinessException(NOT_FOUND_CART));
+            .orElseThrow(() -> new BusinessException(NOT_FOUND_CART));
     }
 
     private CartResponse getCartResponse(Cart cart) {
         return new CartResponse(
-                cart.getCartId(),
-                cart.getItem().getItemId(),
-                cart.getItem().getItemName(),
-                cart.getItem().getItemPrice(),
-                cart.getCount(),
-                cart.getItem().getItemState(),
-                cart.getItem().getItemOption().getOptionValues().stream()
-                        .map(option -> new CartResponse.Option(option.key(), option.value()))
-                        .toList()
+            cart.getCartId(),
+            cart.getItem().getItemId(),
+            cart.getItem().getItemName(),
+            cart.getItem().getItemPrice(),
+            cart.getCount(),
+            cart.getItem().getItemState(),
+            cart.getItem().getItemOption().getOptionValues().stream()
+                .map(option -> new CartResponse.Option(option.key(), option.value()))
+                .toList()
         );
     }
 }
